Allows winner to be determined
Allows winner to be determined by replacing the status declaration in
Board's render with conditional statement. HandleClick in Board  is
changed to return early and ignore click if someone has already won the
or if a square is already filled. Function at end of file has array of
winning states and compares current board to array.